requirements.txt
fastplot
pandas<2.0
pyyaml
rich
scipy
typing-extensions
